fix: Rework database and watchdog to start up more securely

+++ src/backend/adapters/sqlite-adapter.ts
@@ -17,29 +17,66 @@ export class SQLiteFile {
 
 @injectable()
 export class SQLiteAdapter implements IDatabaseAdapter {
-  readonly database: Database
+  database?: Database
 
-  constructor(@inject('Config') private config: IConfig) {
-    this.database = new Database('database.sqlite', error => {
-      if (error) {
-        throw Error(`Unable to create SQLite Database. ${error.message}`)
+  constructor(@inject('Config') private config: IConfig) { }
+
+  open(): Promise<void> {
+    return new Promise<void>((resolve, reject) => {
+      this.database = new Database('database.sqlite', async error => {
+        if (error) {
+          reject(Error(`Unable to create SQLite Database. ${error.message}`))
+        }
+        else {
+          await new Promise(resolve => {
+            this.database!.run('CREATE TABLE IF NOT EXISTS files (terminationTime INTEGER, filename TEXT)', error => {
+              if (error) {
+                console.log(error)
+              }
+
+              resolve()
+            })
+          })
+
+          resolve()
+        }
+      })
+    })
+  }
+
+  close(): Promise<void> {
+    return new Promise<void>(resolve => {
+      if (this.database !== undefined) {
+        this.database.close(error => {
+          if (error) {
+            console.log(error.message)
+          }
+
+          resolve()
+        })
+      }
+      else {
+        resolve()
       }
     })
-
-    this.database.run('CREATE TABLE IF NOT EXISTS files (terminationTime INTEGER, filename TEXT)')
   }
 
   addFile(file: IFile): Promise<void> {
     const sqliteFile = new SQLiteFile(file)
 
     return new Promise<void>(resolve => {
-      this.database.run(`INSERT INTO files VALUES(${sqliteFile.terminationTime}, ?)`, sqliteFile.filename, error => {
-        if (error) {
-          console.log(error.message)
-        }
+      if (this.database !== undefined) {
+        this.database.run(`INSERT INTO files VALUES(${sqliteFile.terminationTime}, ?)`, sqliteFile.filename, error => {
+          if (error) {
+            console.log(error.message)
+          }
 
+          resolve()
+        })
+      }
+      else {
         resolve()
-      })
+      }
     })
   }
 
@@ -47,40 +84,33 @@ export class SQLiteAdapter implements IDatabaseAdapter {
     const now = new Date().getTime()
 
     return new Promise<void>(resolve => {
-      this.database.all(`SELECT * FROM files WHERE terminationTime < ${now}`, async (error, files: SQLiteFile[]) => {
-        if (error) {
-          console.log(error.message)
-        }
-        else {
-          for (const file of files) {
-            fs.unlink(path.join(this.config.uploadDir, file.filename), () => null)
+      if (this.database !== undefined) {
+        this.database.all(`SELECT * FROM files WHERE terminationTime < ${now}`, async (error, files: SQLiteFile[]) => {
+          if (error) {
+            console.log(error.message)
+          }
+          else {
+            for (const file of files) {
+              fs.unlink(path.join(this.config.uploadDir, file.filename), () => null)
+            }
+
+            await new Promise(resolve => {
+              this.database!.run(`DELETE FROM files WHERE terminationTime < ${now}`, error => {
+                if (error) {
+                  console.log(error.message)
+                }
+
+                resolve()
+              })
+            })
           }
 
-          await new Promise(resolve => {
-            this.database.run(`DELETE FROM files WHERE terminationTime < ${now}`, error => {
-              if (error) {
-                console.log(error.message)
-              }
-
-              resolve()
-            })
-          })
-        }
-
+          resolve()
+        })
+      }
+      else {
         resolve()
-      })
-    })
-  }
-
-  close(): Promise<void> {
-    return new Promise<void>(resolve => {
-      this.database.close(error => {
-        if (error) {
-          console.log(error.message)
-        }
-
-        resolve()
-      })
+      }
     })
   }
 }
